refactor: move worker init logic from main.py to workers/ module
Extract init_llm_worker(), build_kv_connector_config(), and
get_engine_runtime_config() into workers/llm_worker.py. Add unified
init_worker() dispatcher in workers/__init__.py that routes by
modality. main.py now only handles CLI entry, signal setup, and
runtime creation.
